requirements.txt
anyio==4.7.0
APScheduler==3.10.4
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
pyngrok==7.2.1
python-telegram-bot==21.8
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3